server/requirements.txt
opencv-python
pytesseract
pdf2image
requests
numpy
pillow
